requirements.txt
streamlit 
Google-generativeai
youtube_transcript_api # python API to allow you to get the transcript for a given youtube video
pathlib
python-dotenv
reportlab